requirements.txt
feedparser
python-telegram-bot==12.8